rename `models/aux.py` -> `auxil.py` to avoid Windows reserved name
