+++ api/models/__init__.py
@@ -1,4 +1,4 @@
-from .aux import InternalModelWithName, MultilingualModel
+from .auxil import InternalModelWithName, MultilingualModel
 from .days_time_slots import DayAndTimeSlot, DayOfWeek, TimeSlot
 from .enrollment_test import (
     EnrollmentTest,
